# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "牧民健康应用API",
    "description": "牧民健康应用的后端API接口",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "timestamp": {
            "type": "integer",
            "description": "时间戳（毫秒）"
          }
        }
      },
...
      "ConsultSession": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID"
          },
          "symptoms": {
            "type": "string",
            "description": "症状描述"
          },
          "duration": {
            "type": "string",
            "description": "症状持续时间"
          },
          "status": {
            "type": "string",
            "description": "会话状态（waiting/consulting/ended）"
          },
          "doctorName": {
            "type": "string",
            "description": "医生姓名"
          },
          "doctorDepartment": {
            "type": "string",
            "description": "医生科室"
          },
          "doctorTitle": {
            "type": "string",
            "description": "医生职称"
          },
          "doctorAvatar": {
            "type": "string",
            "description": "医生头像URL"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time",
            "description": "最后活动时间"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "ConsultMessage": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "string",
            "description": "消息ID"
          },
          "content": {
            "type": "string",
            "description": "消息内容"
          },
          "contentType": {
            "type": "string",
            "description": "内容类型（text/image/audio）"
          },
          "sender": {
            "type": "string",
            "description": "发送者（user/doctor/ai）"
          },
          "hasRead": {
            "type": "boolean",
            "description": "是否已读"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/consult/sessions": {
      "get": {
        "tags": ["问诊服务"],
        "summary": "获取问诊会话列表",
        "description": "获取用户的问诊会话列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "会话状态（waiting/consulting/ended）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer",
                              "description": "总记录数"
                            },
                            "pages": {
                              "type": "integer",
                              "description": "总页数"
                            },
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/ConsultSession"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          }
        }
      },
      "post": {
        "tags": ["问诊服务"],
        "summary": "创建问诊会话",
        "description": "创建新的问诊会话",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["symptoms"],
                "properties": {
                  "symptoms": {
                    "type": "string",
                    "description": "症状描述"
                  },
                  "duration": {
                    "type": "string",
                    "description": "症状持续时间"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "sessionId": {
                              "type": "string",
                              "description": "会话ID"
                            },
                            "createTime": {
                              "type": "string",
                              "format": "date-time",
                              "description": "创建时间"
                            },
                            "status": {
                              "type": "string",
                              "description": "会话状态"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "认证失败"
          }
        }
      }
    },
    "/consult/sessions/{session_id}": {
      "get": {
        "tags": ["问诊服务"],
        "summary": "获取问诊会话详情",
        "description": "获取指定会话ID的问诊会话详情",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "description": "会话ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ConsultSession"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          },
          "404": {
            "description": "会话不存在"
          }
        }
      }
    },
    "/consult/sessions/{session_id}/messages": {
      "get": {
        "tags": ["问诊服务"],
        "summary": "获取问诊消息记录",
        "description": "获取指定会话的消息记录",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "description": "会话ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "查询此时间之前的消息（ISO格式时间戳）",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回消息数量限制",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "hasMore": {
                              "type": "boolean",
                              "description": "是否有更多消息"
                            },
                            "messages": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/ConsultMessage"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          },
          "404": {
            "description": "会话不存在"
          }
        }
      },
      "post": {
        "tags": ["问诊服务"],
        "summary": "发送问诊消息",
        "description": "在指定会话中发送问诊消息",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "description": "会话ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "消息内容"
                  },
                  "content_type": {
                    "type": "string",
                    "description": "内容类型（text/image/audio）",
                    "default": "text"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "messageId": {
                              "type": "string",
                              "description": "消息ID"
                            },
                            "content": {
                              "type": "string",
                              "description": "消息内容"
                            },
                            "contentType": {
                              "type": "string",
                              "description": "内容类型"
                            },
                            "sender": {
                              "type": "string",
                              "description": "发送者"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time",
                              "description": "创建时间"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "认证失败"
          },
          "404": {
            "description": "会话不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  }
} 